Remove explicit reference to XDG_CONFIG_HOME path.
Although if a user sets XDG_CONFIG_HOME it's likely they'll interpret
the documentation correctly, it's perhaps better not to explicitly
refer to the ~/.config directory; "hardcoding" the path in the
documentation seems counter to the flexibility of the XDG Base
Directory Specification. Here, the files are referred to as "system"
and "user" configuration files with an extended explanation of where
these are in the man page.